+++ CHANGELOG.md
@@ -16,6 +16,8 @@ v0.10.0 (in development)
       use `digests` instead
     - Digest fragments are now removed from `DistributionPackage.url` when
       parsing HTML responses
+- Warn on encountering a repository version with a greater minor version than
+  expected
 
 v0.9.0 (2021-08-26)
 -------------------