+++ docs/changelog.rst
@@ -27,6 +27,9 @@ v0.10.0 (in development)
   - Digest fragments are now removed from `DistributionPackage.url` when
     parsing HTML responses
 
+- Warn on encountering a repository version with a greater minor version than
+  expected
+
 
 v0.9.0 (2021-08-26)
 -------------------